requirements.txt
litellm>=1.30.0
tiktoken>=0.6.0
scikit-learn>=1.4.0
pyyaml>=6.0
rich>=13.0
click>=8.1
